requirements.txt
Scrapy==2.11.1
scrapy-poet==0.24.0
scrapy-zyte-api[provider] @ git+https://github.com/ptsonev/scrapy-zyte-api.git
scrapy-spider-metadata
pydantic~=2.9.2
attrs~=24.2.0
dateparser
yarl
form2request @ git+https://github.com/ptsonev/form2request.git
Jinja2
scrapinghub[msgpack]
spidermon @ git+https://github.com/ptsonev/spidermon#egg=spidermon
premailer
SQLAlchemy
pymysql
tldextract
feedparser
pytz
beautifulsoup4
tenacity
jsonpath-ng
